Adapt ClumpMaterialAsset
